+++ packages/apollo-client/src/core/ObservableQuery.ts
@@ -433,7 +433,7 @@ export class ObservableQuery<T> extends Observable<ApolloQueryResult<T>> {
     // since setVariables restarts the subscription, we reset the tornDown status
     this.isTornDown = false;
 
-    const newVariables = variables
+    const newVariables = variables ? variables : this.variables;
 
     if (isEqual(newVariables, this.variables) && !tryFetch) {
       // If we have no observers, then we don't actually want to make a network