+++ packages/apollo-client/src/core/ObservableQuery.ts
@@ -244,8 +244,9 @@ export class ObservableQuery<
   }
 
   public refetch(variables?: TVariables): Promise<ApolloQueryResult<TData>> {
+    const { fetchPolicy } = this.options;
     // early return if trying to read from cache during refetch
-    if (this.options.fetchPolicy === 'cache-only') {
+    if (fetchPolicy === 'cache-only') {
       return Promise.reject(
         new Error(
           'cache-only fetchPolicy option should not be used together with query refetch.',
@@ -268,9 +269,13 @@ export class ObservableQuery<
     }
 
     // Override fetchPolicy for this call only
+    // only network-only and no-cache are safe to use
+    const isNetworkFetchPolicy =
+      fetchPolicy === 'network-only' || fetchPolicy === 'no-cache';
+
     const combinedOptions: WatchQueryOptions = {
       ...this.options,
-      fetchPolicy: 'network-only',
+      fetchPolicy: isNetworkFetchPolicy ? fetchPolicy : 'network-only',
     };
 
     return this.queryManager
@@ -437,7 +442,7 @@ export class ObservableQuery<
     // since setVariables restarts the subscription, we reset the tornDown status
     this.isTornDown = false;
 
-    const newVariables = Object.assign({}, this.variables, variables);
+    const newVariables = variables ? variables : this.variables;
 
     if (isEqual(newVariables, this.variables) && !tryFetch) {
       // If we have no observers, then we don't actually want to make a network