+++ packages/graphql-anywhere/src/utilities.ts
@@ -103,7 +103,9 @@ export function propType(doc) {
   return createChainableTypeChecker((props, propName) => {
     const prop = props[propName];
     try {
-      check(doc, prop);
+      if (!prop.loading) {
+        check(doc, prop);
+      }
       return null;
     } catch (e) {
       // Need a much better error.