+++ packages/apollo-client/CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ### vNEXT
 
+- fixed edge case bug of changing fetchPolicies right after resetStore with no variables present
+
 ### 2.2.8
 - Added the graphQLResultHasError in QueryManager.ts to check not only if result.errors is null, but also empty.
 - Errors occurring during fetchMore no longer errors out the original query [PR#2906](https://github.com/apollographql/apollo-client/pull/2906)