+++ packages/apollo-client/CHANGELOG.md
@@ -10,6 +10,7 @@
 - onResetStore callbacks occur before refetching Observable Queries[PR#3010](https://github.com/apollographql/apollo-client/pull/3010)
 - Error message for in flight queries during `resetStore` includes link completion note[PR#3010](https://github.com/apollographql/apollo-client/pull/3010)
 - Fix navigator being undefined on React Native [PR##3164](https://github.com/apollographql/apollo-client/pull/3164)
+- Typescript improvements. Made observable query parameterized on data and variables: `ObservableQuery<TData, TVariables>` [PR#3140](https://github.com/apollographql/apollo-client/pull/3140)
 
 ### 2.2.3
 - dependency updates