+++ packages/apollo-client/CHANGELOG.md
@@ -2,14 +2,25 @@
 # Change log
 
 ### vNEXT
+
+- fixed edge case bug of changing fetchPolicies right after resetStore with no variables present
+
+### 2.2.8
+- Added the graphQLResultHasError in QueryManager.ts to check not only if result.errors is null, but also empty.
+- Errors occurring during fetchMore no longer errors out the original query [PR#2906](https://github.com/apollographql/apollo-client/pull/2906)
 - Map coverage to original source
+
+### 2.2.4
 - Added `getCacheKey` function to the link context for use in state-link [PR#2998](https://github.com/apollographql/apollo-client/pull/2998)
+- Fix `no-cache` fetch policy so it returns data [PR#3102](https://github.com/apollographql/apollo-client/pull/3102)
 - Fix Memory Leak in Query Manager [PR#3119](https://github.com/apollographql/apollo-client/pull/3119)
 - Pass non-optimistic query to `subscribeToMore` updateQuery
 [PR#3068](https://github.com/apollographql/apollo-client/pull/3068)
 - onResetStore callbacks occur before refetching Observable Queries[PR#3010](https://github.com/apollographql/apollo-client/pull/3010)
 - Error message for in flight queries during `resetStore` includes link completion note[PR#3010](https://github.com/apollographql/apollo-client/pull/3010)
+- `ApolloError` can now be checked with `instanceof` operator 
 - Fix navigator being undefined on React Native [PR##3164](https://github.com/apollographql/apollo-client/pull/3164)
+- Remove spread of variables preventing issues with removing keys [#3081](https://github.com/apollographql/apollo-client/pull/3081)
 - Typescript improvements. Made observable query parameterized on data and variables: `ObservableQuery<TData, TVariables>` [PR#3140](https://github.com/apollographql/apollo-client/pull/3140)
 
 ### 2.2.3